Broken game objects after the merge w/ Develop
