+++ Assets/Scripts/UI/HighScoreAdder.cs
