+++ Assets/Scripts/_Gameplay/Drinks/ThreeDDrinkObject.cs.meta
